Splitting stack trace support out of status.c.
