+++ runtime/src/iree/base/BUILD.bazel
@@ -33,6 +33,8 @@ iree_runtime_cc_library(
         "status.c",
         "status.h",
         "status_cc.h",
+        "status_payload.h",
+        "status_stack_trace.c",
         "string_builder.c",
         "string_builder.h",
         "string_view.c",
